Move nested_get to Features class
