+++ connectors/tests/test_byoc.py
@@ -906,6 +906,33 @@ def test_sync_rules_enabled(features_json, sync_rules_enabled):
     assert features.sync_rules_enabled() == sync_rules_enabled
 
 
+@pytest.mark.parametrize(
+    "nested_dict, keys, default, expected",
+    [
+        # extract True
+        ({"a": {"b": {"c": True}}}, ["a", "b", "c"], False, True),
+        (
+            {"a": {"b": {"c": True}}},
+            # "d" doesn't exist -> fall back to False
+            ["a", "b", "c", "d"],
+            False,
+            False,
+        ),
+        (
+            {"a": {"b": {"c": True}}},
+            # "wrong_key" doesn't exist -> fall back to False
+            ["wrong_key", "b", "c"],
+            False,
+            False,
+        ),
+        # fallback to True
+        (None, ["a", "b", "c"], True, True),
+    ],
+)
+def test_nested_get(nested_dict, keys, default, expected):
+    assert expected == Features(nested_dict)._nested_feature_enabled(keys, default)
+
+
 def test_pending_job_query_with_connectors_ids(mock_responses, set_env):
     config = load_config(CONFIG)
